Updated libopus to 1.2.1.
1.2.1

This Opus 1.2.1 minor release fixes a relatively rare issue where
the 1.2 encoder would wrongly assume a signal to be bandlimited to
12 kHz and not encode frequencies between 12 and 20 kHz. This only
happens on a few clips, but it is good to update to avoid a potential
loss of quality.

1.2.

This Opus 1.2 major release brings many quality improvements, new
features, and bug fixes. You can read all the details in this release
demo page. Changes since 1.1.x include:


    Speech quality improvements especially in the 12-20 kbit/s range
    Improved VBR encoding for hybrid mode
    More aggressive use of wider speech bandwidth, including fullband speech starting at 14 kbit/s
    Music quality improvements in the 32-48 kb/s range
    Generic and SSE CELT optimizations
    Support for directly encoding packets up to 120 ms
    DTX support for CELT mode
    SILK CBR improvements
    Support for all of the fixes in draft-ietf-codec-opus-update-06 (the mono downmix and the folding fixes need --enable-update-draft)
    Many bug fixes, including integer wrap-arounds discovered through fuzzing (no security implications)