textproc/ja-sed: Fix empty installed GNU info file.
This package applies a vendor patch against the distributed GNU
sed files to extend GNU sed with multibyte awareness.  However, the
changes to the GNU info file are applied before the changes to its
source file, so the timestamps make it appear that the GNU info
file needs to be rebuilt.

Update the timestamp of the patched sed.info file in a "pre-build"
target to correct this issue.

Bump the PKGREVISION due to a change in the installed GNU info
file.