Update tex-graphics{,-doc} to 1.2c
This release features a major overhaul of
the driver (.def) files, bringing them back
into line both with each other and with the
core LaTeX graphics package.

As-of this release, new graphics features are
incorporated into graphicx, and the .def files
are returned to providing support for these
features but without adding additional keys.

The following keys are added to the
officially-supported set:
- "interpolate" (bitmap images)
- "page" (PDF images)
- "pagebox" (all images)
- "quiet" (all images)

The internal structures of the various .def
files have also been aligned based around
dvips.def, and some support material has been
moved to graphicx.